decouple trainer from parallelwrapper itself, this allows for parameter server impl
