+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/impl/multilayer/SparkDl4jMultiLayer.java
@@ -18,7 +18,6 @@
 
 package org.deeplearning4j.spark.impl.multilayer;
 
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.SparkContext;
 import org.apache.spark.api.java.JavaDoubleRDD;
@@ -29,20 +28,16 @@ import org.apache.spark.mllib.linalg.Matrix;
 import org.apache.spark.mllib.linalg.Vector;
 import org.apache.spark.mllib.regression.LabeledPoint;
 import org.apache.spark.rdd.RDD;
-import org.deeplearning4j.api.storage.StatsStorageRouter;
-import org.deeplearning4j.api.storage.StatsStorageRouterProvider;
-import org.deeplearning4j.api.storage.listener.RoutingIterationListener;
 import org.deeplearning4j.eval.*;
 import org.deeplearning4j.nn.conf.MultiLayerConfiguration;
 import org.deeplearning4j.nn.conf.layers.FeedForwardLayer;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
-import org.deeplearning4j.optimize.api.IterationListener;
 import org.deeplearning4j.spark.api.TrainingMaster;
 import org.deeplearning4j.spark.api.stats.SparkTrainingStats;
 import org.deeplearning4j.spark.impl.SparkListenable;
 import org.deeplearning4j.spark.impl.common.reduce.IntDoubleReduceFunction;
-import org.deeplearning4j.spark.impl.listeners.VanillaStatsStorageRouterProvider;
-import org.deeplearning4j.spark.impl.multilayer.evaluation.*;
+import org.deeplearning4j.spark.impl.multilayer.evaluation.IEvaluateFlatMapFunction;
+import org.deeplearning4j.spark.impl.multilayer.evaluation.IEvaluationReduceFunction;
 import org.deeplearning4j.spark.impl.multilayer.scoring.FeedForwardWithKeyFunction;
 import org.deeplearning4j.spark.impl.multilayer.scoring.ScoreExamplesFunction;
 import org.deeplearning4j.spark.impl.multilayer.scoring.ScoreExamplesWithKeyFunction;
@@ -63,8 +58,7 @@ import scala.Tuple2;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.Serializable;
-import java.util.*;
+import java.util.List;
 
 /**
  * Master class for spark
@@ -367,7 +361,8 @@ public class SparkDl4jMultiLayer extends SparkListenable {
     }
 
     /**
-     * {@code RDD<DataSet>} overload of {@link #scoreExamples(JavaRDD, boolean, int)}
+     * {@code RDD<DataSet>}
+     * overload of {@link #scoreExamples(JavaRDD, boolean, int)}
      */
     public JavaDoubleRDD scoreExamples(RDD<DataSet> data, boolean includeRegularizationTerms, int batchSize) {
         return scoreExamples(data.toJavaRDD(), includeRegularizationTerms, batchSize);
